Reorganized UI scripts into a better organization
