+++ Assets/Prefabs/UI/Components/GraphicsPanel.prefab
@@ -486,7 +486,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &4536500174311316358
 RectTransform:
   m_ObjectHideFlags: 0
@@ -5399,10 +5399,10 @@ RectTransform:
   m_Father: {fileID: 5316467488556632297}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 149.99998}
-  m_SizeDelta: {x: 500, y: 400}
+  m_AnchorMin: {x: 0.5, y: 1}
+  m_AnchorMax: {x: 0.5, y: 1}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 500, y: 0}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!114 &1479094264272974124
 MonoBehaviour:
@@ -5736,8 +5736,8 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 5840620962272616998}
   m_HandleRect: {fileID: 4767489105610624174}
   m_Direction: 2
-  m_Value: 0.9999998
-  m_Size: 0.75
+  m_Value: 0
+  m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
