+++ Assets/Prefabs/UI/Components/SoundPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ad55bbb5205d75443a0ccd040e3f522d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
