+++ Assets/Scripts/UI/Chat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d3368ea0be6a9945af2e10042f411a7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
