+++ Assets/Scripts/UI/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df16d5925f8d892438d441fa85e7ceac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
