+++ Assets/Scripts/UI/Events/CursorStateOnMenuLoad.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace PropHunt.UI
+namespace PropHunt.UI.Events
 {
     /// <summary>
     /// Simple class to set cursor state when the menu loads
