+++ Assets/Tests/EditMode/UI/ChatBoxAutoScrollTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections;
 using NUnit.Framework;
 using PropHunt.Game.Communication;
-using PropHunt.UI;
+using PropHunt.UI.Chat;
 using UnityEngine;
 using UnityEngine.TestTools;
 using UnityEngine.UI;
